Display difficulty in skill roll output
Shows the difficulty of the roll as part of the 'Sucessess'
output on the skill roll, so it is clear what difficulty you were
rolling against.

Also I changed the default for difficulty to be 0 if it is omitted
on the command line, if this should not be possible then the command
parsing should be checked over.